+++ src/impl/regexParser.ts
@@ -234,7 +234,7 @@ function extractRFC2822(match: RegExpExecArray): ParseResult {
 function preprocessRFC2822(s: string) {
     // Remove comments and folding whitespace and replace multiple-spaces with a single space
     return s
-        .replace(/\([^)]*\)|[\n\t]/g, " ")
+        .replace(/\([^()]*\)|[\n\t]/g, " ")
         .replace(/(\s\s+)/g, " ")
         .trim();
 }