+++ test/datetime/create.test.ts
@@ -616,7 +616,7 @@ test("DateTime.fromObject accepts a locale", () => {
 
 test("DateTime.fromObject accepts a locale with calendar and numbering identifiers", () => {
     const res = DateTime.fromObject({}, { locale: "be-u-ca-coptic-nu-mong" });
-    expect(res.locale).toBe("be");
+    expect(res.locale).toBe("be-u-ca-coptic-nu-mong");
     expect(res.outputCalendar).toBe("coptic");
     expect(res.numberingSystem).toBe("mong");
 });
@@ -630,7 +630,7 @@ test("DateTime.fromObject accepts a locale string with weird junk in it", () => 
             }
         );
 
-        expect(res.locale).toBe("be");
+        expect(res.locale).toBe("be-u-ca-coptic-ca-islamic");
 
         // "coptic" is right, but some versions of Node 10 give "gregory"
         expect(res.outputCalendar === "gregory" || res.outputCalendar === "coptic").toBe(true);
@@ -648,7 +648,7 @@ test("DateTime.fromObject overrides the locale string with explicit settings", (
         }
     );
 
-    expect(res.locale).toBe("be");
+    expect(res.locale).toBe("be-u-ca-coptic-nu-mong");
     expect(res.outputCalendar).toBe("islamic");
     expect(res.numberingSystem).toBe("thai");
 });
@@ -763,4 +763,19 @@ test("DateTime.fromHTTP is invalid when weekday is not consistent", () => {
 
 test("DateTime.fromObject takes a undefined to mean {}", () => {
     expect(DateTime.fromObject().year).toBe(new Date().getFullYear());
+});
+
+test("private language subtags don't break unicode subtags", () => {
+    const res = DateTime.fromObject(
+        {},
+        {
+            locale: "be-u-ca-coptic-nu-mong-x-twain",
+            numberingSystem: "thai",
+            outputCalendar: "islamic",
+        }
+    );
+
+    expect(res.locale).toBe("be-u-ca-coptic-nu-mong");
+    expect(res.outputCalendar).toBe("islamic");
+    expect(res.numberingSystem).toBe("thai");
 });